.kamila/llm/requirements_gemini.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.0
structlog>=23.2.0
